Change the ESC control logic to PWM in microseconds
In lieu of 0 degrees being the minimum, the ESC has been set 1.1ms for
the neutral position, and instead of 180 being the maximum, 2.0ms has
been set for full forward power. As such, the control logic needs to be
updated to accept these new values and the minimum and maximum.